data-generator/requirements.txt
Faker==22.5.1
psycopg2-binary==2.9.9
pandas==2.2.0
sqlalchemy==2.0.25